+++ packages/frontend/src/components/message/MessageListAndComposer.tsx
@@ -2,7 +2,8 @@ import React, { useRef, useEffect, useCallback } from 'react'
 import { join, parse, ParsedPath } from 'path'
 import { T } from '@deltachat/jsonrpc-client'
 
-import Composer, { useDraft } from '../composer/Composer'
+import Composer from '../composer/Composer'
+import { useDraft } from '../../hooks/chat/useDraft'
 import { getLogger } from '../../../../shared/logger'
 import MessageList from './MessageList'
 import type ComposerMessageInput from '../composer/ComposerMessageInput'
@@ -15,6 +16,7 @@ import { ReactionsBarProvider } from '../ReactionsBar'
 import useDialog from '../../hooks/dialog/useDialog'
 import useMessage from '../../hooks/chat/useMessage'
 import { Viewtype } from '@deltachat/jsonrpc-client/dist/generated/types'
+import { useMessageList } from '../../stores/messagelist'
 
 const log = getLogger('renderer/MessageListAndComposer')
 
@@ -100,6 +102,14 @@ export default function MessageListAndComposer({ accountId, chat }: Props) {
 
   const regularMessageInputRef = useRef<ComposerMessageInput>(null)
   const editMessageInputRef = useRef<ComposerMessageInput>(null)
+
+  const {
+    store: messageListStore,
+    state: messageListState,
+    fetchMoreBottom,
+    fetchMoreTop,
+  } = useMessageList(accountId, chat.id)
+
   const {
     draftState,
     updateDraftText,
@@ -111,6 +121,7 @@ export default function MessageListAndComposer({ accountId, chat }: Props) {
     clearDraftStateAndUpdateTextareaValue,
     setDraftStateAndUpdateTextareaValue,
   } = useDraft(
+    messageListState,
     accountId,
     chat.id,
     chat.isContactRequest,
@@ -294,6 +305,10 @@ export default function MessageListAndComposer({ accountId, chat }: Props) {
               accountId={accountId}
               chat={chat}
               refComposer={refComposer}
+              messageListStore={messageListStore}
+              messageListState={messageListState}
+              fetchMoreBottom={fetchMoreBottom}
+              fetchMoreTop={fetchMoreTop}
             />
           </ReactionsBarProvider>
         </RecoverableCrashScreen>